# Optimize pwndbg.commands.ai import time